ACDC/requirements.txt
numpy==1.23.5
pandas==1.5.2
seaborn==0.13.2
matplotlib==3.7.1
scikit-learn==1.2.0
umap-learn==0.5.7
scipy==1.10.1
phenograph==1.5.7
fcsy==0.10.0
flowio==1.3.0
plotly==5.24.1
packaging==23.1
tenacity==6.2.0
scanpy==1.10.3